requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langgraph>=0.0.10
gradio>=4.0.0
duckduckgo-search>=4.1.1
requests>=2.31.0
python-dotenv>=1.0.0 